refactor(builder): modularize build and prepare logic
- Moved build logic to `src/builder/build/mod.rs`
- Moved prepare logic to `src/builder/prepare/mod.rs`
- Updated `src/builder/service.rs` to use the new modularized functions
- Added `mod build` and `mod prepare` to `src/builder/mod.rs` for module inclusion
- Simplified `impl PackageService for Packager` by delegating to the new modularized functions
